fix() types split
