+++ packages/types/src/TutimOptions.ts
@@ -1,0 +1,18 @@
+import { FieldComponents } from './Field';
+import { FormConfig } from './FormConfig';
+
+export interface TutimOptions {
+  forms: Record<string, FormConfig>;
+  clientId?: string;
+}
+
+export interface TutimOptionsProviderValue {
+  options: TutimOptions;
+  setOptions: (config: TutimOptions) => void;
+}
+
+export interface FormProviderProps {
+  fieldComponents?: FieldComponents;
+  options?: Partial<TutimOptions>;
+  children: React.ReactNode;
+}
